[core] [sky] refs #189 - Move SkyAPI interface src/core => src/coin/skycoin/skytypes
